reorganized files, separating styled-components and components
