--- client/components/DropdownMenu.tsx
@@ -1,11 +1,0 @@
-import React, { useState } from 'react';
-import Button from './Button';
-
-function DropdownMenu() {
-
-    return (
-        <div className='dropdown'>
-
-        </div>
-    )
-}
