--- client/components/MenuBtn.tsx
@@ -1,9 +1,0 @@
-import React, { useState } from 'react';
-import Button from './Button';
-
-
-function MenuBtn() {
-  const [isOpened, setIsOpened] = useState<boolean>(false);
-
-  return <Button icon={} iconClass={} btnClass={} onClickFunc={() => {setIsOpened((prev) => !prev)}} />;
-}
